+++ docs/changelog.md
@@ -2,6 +2,11 @@
 
 _Copyright (c) 2013 Daniele Veneroni. Released under MIT License._
 
+### Update 25/05/2013
+
+* Improved random map generation
+* Bugfix
+
 ### Update 24/05/2013
 
 * Add new units: Catapult, Horseman, Galleon, Knight, Spearman, Swordsman, Trireme, Lancer